# perf: Improve Xor method performance by ~20% for big sets